Separate Map from Dashboard
