+++ src/components/Navbar/styles.scss
@@ -4,6 +4,7 @@
     justify-content: space-between;
     background-color: var(--color-navbar);
     padding: 0 calc(var(--spacing-large) - var(--spacing-medium));
+    height: var(--height-navbar);
 
     .brand {
         display: flex;
